"fix the app crash due to non-numeric amount issue in add entries activities"

+++ app/src/main/java/com/example/budgetapp/Expense.kt
@@ -4,6 +4,6 @@ import java.io.Serializable
 
 data class Expense (var id: Int? = null,
                     var categories: String = "",
-                    var percentage: Int = 0,
-                    var max: Int = 0
+                    var percentage: Double = 0.0,
+                    var max: Double = 0.0
                     ): Serializable
